st-hal: Fix concurrency issue for non-dedicated path
If concurrency is not enabled and dedicated path is not being used,
the sound trigger session should be paused while audio capture is
active. After such pause, the LPI transition was resuming the
session which is erroneous behavior. This change also updates the
policy for moving to Non-LPI due to concurrency. For audio capture
usecases, the lower layers will exit island mode but the sound
trigger session will not transition to Non-LPI. Audio playback,
voice call, and voip call will all transition the sound trigger
sessions to Non-LPI and enable EC.