st-hal: Always reset EC ref as part of the disable usecase sequence
The reference count for echo reference is being handled in driver
layer, so in sthal each session can set and reset echo reference
independently. The issue was seen when 2 VA sessions were active
while playback was active on bluetooth headset. When a wired
headset was connected, the echo reference had to switch paths.
Previously the echo reference could only be reset from sthal once,
so the driver reference count never got back to 0 during the switch
which caused an error.

Change-Id: I45228ef4cb8aedea4c78bd349efb6931886bdc6b