st-hal: fix session back to active when disabling capture usecase
When va concurrency is not allowed for audio capture, if there
are two concurrent active capture usecases and one of them gets
disabled, but va concurrency is still not allowed, then va
session should keep in paused state.
Also, unexpected pause and resume can break the path of
remaining audio capture usecase.
Remove the check of cached concurrency flag to fix it.

Change-Id: I9b64e0c2faab37b0bcc78288de7516cf8996d66f