st-hal: fix st session potential deadlock
In function hw_sess_cb, st_ses->lock could be left as unlocked
in switch case ST_HW_SESS_EVENT_DETECTED when "!lock_status"
condition is met and the session is not in active state. Due to
the switch case exits prematurely, st_ses->lock will not be
unlocked under this condition. Remove the "break" to avoid this
scenario in function hw_sess_cb.

Change-Id: I68af8bc46334a71325fb25fc101b9950c207b2e3