st-hal: Fix module params assignment in lsm_usecase
A single lsm_usecase is used for separate hw sessions for
GMM and PDK5 sound models. Currently, the module params in the
lsm_usecase are shared and set only during load sound model
sequence. This can lead to an issue if GMM is loaded, then PDK5
is loaded, then GMM is started because the module params were
overwritten by the other sound model type. This change provides
a separate lsm_usecase for each hw session so that the data is no
longer shared and overwritten.

Change-Id: I054a73a0ba1e669930b6a6f1a76d5b9e9e316797