st-hal: Initialize hw properties on each stdev_open call
Currently, the hw properties are only initialized when the
stdev_get_properties or stdev_get_properties_extended APIs are
called. In audioserver restart usecases, the framework uses
cached hw properties values so it does not call the get properties
API again. This means sthal does not have the initialized values.

Change-Id: I8f2259bfa90be0b238d2c07725ffb7b7fe7faad0