st-hal: Prevent erroneous detection event dispatch
Currently, the state functions assign the client session from
the event params. Detection events do not have the client session
because it is unknown at the time of dispatch. This can cause
NULL pointer dereference when non active state functions receive
a detection event. This change only allows detection events to be
sent when the session is in active state.

Change-Id: If34c3e886e3ff6035e70ad8cea6c67548b3fd5ea