st-hal: Prevent erroneous detections while client is buffering
Pause events will interrupt buffering in sthal without notifying
the client. Thus, the client will continue buffering anyway until
it receives the requested amount of data. In this state, there
could be a resume event and then the session could make more
detections. Calling the client callback while the client is still
buffering causes errors in the middleware state machine, so this
scenario is prevented with this change. The resume events in this
usecase will now be ignored, and the session will only start again
when the client calls start recognition.

Change-Id: I2b115c980b836f154dad23f69171e4c18ba30d20