st-hal: Avoid deadlock in aggregator thread loop
The potential deadlock scenario is as follows - the aggregator
thread loop always has the ss_detections_lock unless it is waiting
for a signal. The aggregator thread loop also takes the
st_session_lock for processing detection events. The client can
call unload sound model, which takes the st_session_lock, followed
by the ss_detections_lock in order to exit the aggregator thread
loop. This potential deadlock is fixed using a trylock.

Change-Id: I262cc88054e92c3487adad39046e5d5014e99beb