hal: Reset the backend when LPI mode is changed due to battery status change.
When SVA state is in buffering and if remove the power cable,
it should change the LPI mode from NLPI to LPI.For that we have to
reset the backend, so that when next detection is triggered it will
have changed backend and both FE and BE will be in LPI mode.
Handle this usecase scenario, by resetting the backend during hw_session
start.

Change-Id: I00bc25913cdc7c48864a99366e07b3d0368dd5a7