st-hal: update detection_sent flag before calling start/restart
When detection happens in active state, detection_sent flag is
set to false and its set to true just before client is notified of
it. Till the next detection for same client this flag will be true.
In case of Dual VA, if concurrency event happens leading to backend
config update, we will post pause events to other sessions where
depending on detection_sent flag, pending_client_start flag will be
set to true. This pending_client_start flag if true, will prevent
session to resume for that client, when we post resume event to
that client.

Change-Id: I858e4b1fbe23bd13b26f59896a814eb647c2ba21